Refactor de tiene_rol_en a gpa_tag
