+++ templates/base.html
@@ -1,5 +1,5 @@
 {% load static %}
-{% load tiene_rol_en %}
+{% load gpa_tags %}
 {% load socialaccount %}
 <!DOCTYPE html>
 <html lang="es">
