+++ templates/historias/archivos.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 {% block title %}Archivos US{% endblock %}
 {% block contenido %}
-{% load tiene_rol_en %}
+{% load gpa_tags %}
 
 <div class="container">
     <div class="row">
