+++ templates/historias/base.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 {% block title %}Historias{% endblock %}
 {% block contenido %}
-{% load tiene_rol_en %}
+{% load gpa_tags %}
 
 <div>
     <style>
