+++ templates/historias/comentarios.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 {% block title %}Comentarios de {{ historia.nombre }}{% endblock %}
 {% block contenido %}
-{% load tiene_rol_en %}
+{% load gpa_tags %}
 
 <div class="container">
     <div class="row">
