+++ templates/home.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 {% block title %}Gestión de Proyectos Ágil{% endblock %}
 {% block contenido %}
-{% load tiene_rol_en %}
+{% load gpa_tags %}
 
 <!-- Redirigir a home si no tiene los permisos para ver esta pantalla -->
 <div class="container">
