+++ templates/proyectos/base.html
@@ -1,6 +1,6 @@
 {% extends "base.html" %}
 {% block contenido %}
-{% load tiene_rol_en %}
+{% load gpa_tags %}
 
 <!-- Redirigir a home si no tiene los permisos para ver esta pantalla -->
 <div class="container">
