+++ templates/proyectos/home.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 {% block title %}Proyecto {{ proyecto.nombre }}{% endblock %}
 {% block contenido %}
-{% load tiene_rol_en %}
+{% load gpa_tags %}
 <div class="d-flex align-items-center justify-content-center h-75 d-inline-block">
     <div class="box text-center">
         <h1>{{ proyecto.nombre }}</h1>
