+++ templates/proyectos/roles_proyecto/roles_proyecto.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 {% block title %}Roles de proyecto{% endblock %}
 {% block contenido %}
-{% load tiene_rol_en %}
+{% load gpa_tags %}
 
 <!-- Redirigir a home si no tiene los permisos para ver esta pantalla -->
 <div class="container">
