+++ templates/tablero/tablero.html
@@ -2,7 +2,7 @@
 {% block title %}Tablero de {{ tipo.nombre }}{% endblock %}
 {% load static %}
 {% block contenido %}
-{% load tiene_rol_en %}
+{% load gpa_tags %}
 <div class="container-fluid pt-3 overflow-auto">
     <h3>{{ tipo.nombre }}</h3>
     <div class="small">{{ tipo.descripcion }}</div>
