+++ templates/tipos-us/base.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 {% block title %}Tipo de historia de usuario{% endblock %}
 {% block contenido %}
-{% load tiene_rol_en %}
+{% load gpa_tags %}
 <div class="container">
     <div class="row">
         <div class="col-md-12">
