+++ templates/usuarios_equipos/equiporoles.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 {% block title %}Gestion de miembros de {{ proyecto.nombre }}{% endblock %}
 {% block contenido %}
-{% load tiene_rol_en %}
+{% load gpa_tags %}
 {% tiene_rol_en_proyecto user "Scrum Master" proyecto.id as es_master %}
 
 <div class="row mt-5">
